Drop now-unnecessary allowHalfOpen TLS error handling fix
This was previously necessary, due to how half-open connections were
handled in the original implementation of httpolyglot. The
implementation has changed quite dramatically since then though, and
nowadays raw sockets, TLS & HTTP are all handled quite independently,
and this doesn't seem to have any useful effect (and for Node 18+, it
sometimes causes crashes when cleaning up HTTP/2 connections!).

As part of removing this, this also adds a test that covers the original
behaviour that motivated it - this test passes with or without this new
line, in both modern and older Node versions (notably including v12, the
version the original bug was reported for).